Ask to delete config data of plugin if uninstalled (#140)
* Enables deleting of config data

When removing a plugin eg.config is scanned for any eg.PersistantData the plugin may have set. and if it finds any prompts the user if they would like to delete this data. If yes the data is marked for deletion. If the user does not save their EventGhost session all of the marks for deletion are removed. it also handles the ImportAll() when the debug level is set from a command line.

* Adds Python formatter

Adds PrettyPythonPrint to eg.Utils.

This is use to format an attribyte name, attribute value to be written to a file. It formats name = value with indent to conform to 80 character line length it will split tuples and lists but also will wrap long strings with brackets and break the string into multiple lines

* Changes eg.Cli.args.configDir default

Changes eg.Cli.configDir default from None to the actual default folder that eventghost normally uses

* Updates eg.Config and eg.PersistentData

Recode of eg.Config and eg.PersistentData

This recode allows for eg.config to be initilized first and the setting of eg.debugLevel second.. by doing this it removes any gap from when eg starts to when the previous debug lebel was set. 

Created eg.PersistentDataBase which handles any information that is to be written to config.py the base class gets set to each class that gets set into eg.config. and each base class handles it's own writing of the data it holds and also handles setting it's self to be deleted

I also made use of the UndoHandler to keep with EG code and made a new subclass  of the UndoHandler to handle when and when not to save a specific set of config data

all plugin data is marked for delete upon initial oload. and then once the plugin has been added to the tree it will remove the mark for delete. this mechanism works best as i do not have to worry about eventghost being in debug mode or not. I also added the 4 core plugins. EventGhost, Window, Mouse, System to not have this delete flag set.

Utilizes eg.Utils.PrettyPythonPrint to format the output of the config file. 

Added checking for new attribute names being added to the config when doing an upgrade. of either EG or with a plugin. if there is a new addition it will add the new attribute automatically.