Fixes bug in wx.Locale.GetSystemLanguage
OK so this is the deal. After a whole lot of digging I have discovered
that in fact the wx.Locale.GetSystemLanguage in fact does not return
the proper id. It actually returns a locale id and not a language id.

In Windows the locale id is derived from the format in
Region and Language. Text formatting has absolutely nothing to do with
the displayed language. This is the kicker. wxLocale has nothing to do
with Locale. It has everything to do with the Language. and on top of it
GetSystemLanguage ever returned the "system" language that would cause
problems. The system language is defined by changing the non unicode
application language located in the Region and Language panel. wxPython
is a unicode program. So that would really mess things up.

I had to create a cross reference between windows LANGID's and what
wxPython uses for numbers. because they were kind enough to use a while
different set of numbers. The big question would be is "Why??".

I now call the Windows API kernel32.GetUserDefaultUILanguage to get
the LANGID which i cross reference using that table i made and then
pass the proper wx.LANGUAGE_* code to wx.Locale.